requirements.txt
Flask==0.10.1
Flask-BrowserID==0.0.4
Flask-Login==0.2.7
Flask-SQLAlchemy==1.0
Jinja2==2.7.1
Mako==0.9.0
MarkupSafe==0.18
SQLAlchemy==0.8.2
Werkzeug==0.9.4
alembic==0.6.0
gdata==2.0.18
itsdangerous==0.23
requests==2.0.1
suds==0.4
sure==1.2.2
wsgiref==0.1.2
PyYAML==3.10
